@@ -1,10 +1,9 @@
 package br.com.cervamania.cervamania.View;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
@@ -17,7 +16,6 @@ import android.widget.RatingBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import br.com.cervamania.cervamania.Controller.AdapterListaComentariosCervejaSelecionada;
 import br.com.cervamania.cervamania.Controller.TarefaInsereClassificacao;
 import br.com.cervamania.cervamania.Model.ClassificacaoCerveja;
 import br.com.cervamania.cervamania.Model.CoresCervejas;
@@ -49,7 +47,7 @@ public class InsereClassificacaoCervejaSelecionadaActivity extends AppCompatActi
 
         savedInstanceState = getIntent().getExtras();
 
-        toolbar = (android.support.v7.widget.Toolbar) findViewById(R.id.toolBarInsereClassificacaoCervejaSelecionadaActivity);
+        toolbar = (androidx.appcompat.widget.Toolbar) findViewById(R.id.toolBarInsereClassificacaoCervejaSelecionadaActivity);
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setHomeButtonEnabled(true);
